Move error classes to their own files so they can be loaded by Zeitwerk
Closes #313
