+++ lib/solid_queue/processes/process_missing_error.rb
@@ -1,0 +1,9 @@
+module SolidQueue
+  module Processes
+    class ProcessMissingError < RuntimeError
+      def initialize
+        super("The process that was running this job no longer exists")
+      end
+    end
+  end
+end
